+++ packages/zero/src/components/textarea/index.ts
@@ -3,5 +3,6 @@ export type {
     TextareaRootProps,
     TextareaLabelProps,
     TextareaTextareaProps,
+    TextareaHandle,
 } from './Textarea.js';
 export { textareaAnatomy } from './anatomy.js';